requirements.txt
numpy==1.23.5
opencv-python==4.6.0.66
pandas==1.5.2
Pillow==9.3.0